Stablizing rust flake nix
 After a bit of testing I've learned, that
the rust-analyzer is better managed by rustup
globally. No need to install for every package.
We also had some comments to remove, and
I learned that lld's speed does not make
up for the fact that without it installed globally,
not being able to use an editor that doesn't
see nix's lld is crippling, so this is what
we'll have for now.